Avoid 2 hangs during shutdown
Unstable system power resulted in hung PRUs in two separate ways that prevented `powerfail` from performing an orderly shutdown:

1. If /SEEK is endlessly asserted, shutdown fails because PRU0 will not proceed until /SEEK goes high again.
2. After glitched external signals, due to disorderly shutdown in an ornery host, PRU1 signals PRU0 `write_done` and is in `write_done_wait` waiting to hear PRU0 is in `STATE_READ`. But PRU0 is in `selected` waiting to hear `STATE_READ_DONE` from PRU1. Deadlock.

In both cases a check for `PRU0_EXIT` in the command register is made, and EXIT is immediately performed should it be  found. Putting a check for EXIT into `glitch` did not avoid the second case, unfortunately, so the `EXIT` check was inlined directly where the deadlock was observed.